toggle-resolution script is broken
Current script takes current mode from `$EEEPC_VAR/lvds-modes` but doesn't regenerate it. So it switches mode only once instead of looping. The best way to get current mode is from `xrandr`, so we need to call it every time instead of caching. Here is a minimal fix for the problem so toggle-resolution works.