Defaults reflect out of the box ACPI handling
The changeset is large just to show that almost everything is handled these days by default even without `acpid`. Notable exceptions are SHE and touchpad.

Of course it is a good idea to test the config on more than one system so feel free to add LGTM if it works for you.